+++ backoffice/src/App.tsx
@@ -5,14 +5,17 @@ import { AuthGateway } from "gateways/auth/AuthGateway";
 import { BeaconsGateway } from "gateways/beacons/BeaconsGateway";
 import { ExportsGateway } from "gateways/exports/ExportsGateway";
 import { UsesGateway } from "gateways/uses/UsesGateway";
-import React, { FunctionComponent } from "react";
+import { FunctionComponent } from "react";
 import {
   HashRouter as Router,
   Route,
   Switch,
   useParams,
 } from "react-router-dom";
-import { CertificateView, LetterView } from "views/CertificateView";
+import {
+  CertificateView,
+  LetterView,
+} from "views/exports/certificates/CertificateView";
 import "./App.scss";
 import { AuthProvider } from "./components/auth/AuthProvider";
 import { AuthenticatedPOSTButton } from "./components/AuthenticatedPOSTButton";
