+++ backoffice/src/views/exports/certificates/certificate.scss
