+++ backoffice/src/views/exports/letters/letter.scss
@@ -70,7 +70,7 @@
 }
 
 @media print {
-  .certificate {
+  .letter {
     margin: 0;
     box-shadow: none;
   }
